refactor: 불필요한 파일제거 및 패키지 변경
