+++ backend/smody/src/main/java/com/woowacourse/smody/auth/GoogleApi.java
@@ -1,4 +1,4 @@
-package com.woowacourse.smody.controller.openapi;
+package com.woowacourse.smody.auth;
 
 import com.woowacourse.smody.dto.GoogleTokenRequest;
 import com.woowacourse.smody.dto.GoogleTokenResponse;
