--- backend/smody/src/main/java/com/woowacourse/smody/webhook/GitHubHookRequest.java
@@ -1,17 +1,0 @@
-package com.woowacourse.smody.webhook;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class GitHubHookRequest {
-
-    private String ref;
-
-    public String parseBranch() {
-        return ref.split("/")[2];
-    }
-}
